♻️ refactor: 페이지들을 폴더 기반 컴포넌트 구조로 리팩토링
