+++ src/pages/Home/Home.jsx
@@ -1,4 +1,4 @@
-import Header from "../components/Header";
+import Header from "../../components/Header";
 
 const Home = () => {
   return (
